Remove "Compare" class; just use LessFunc directly
As discussed in issue #24, the implemetnations of Compare::le()
and Compare::gt() were more expensive than they needed to be.
As long as LessFunction provides the strict weak ordering that
std::sort() requires, we can safely synthesize all four
lt/gt/le/gt options with a single call.

While the simple fix would be to just change their definitions, I
went a bit further and removed the Compare class entirely.  I don't
really think that it provides much, but it does take space in the
TimSort object whether it's needed or not -- this could be addressed
by some application of the Empty Base Class optimization, but it's
a bit awkward)

I think it was probably done with the intention that having the
comparator as a member of "this" for the core algorithm would be
better because it would cut down on the number of function parameters.
Hoever, the normal way that the STL's sorting algorithms are implemented
is to just pass the comparator down as a parameter to the functions
so that the optimizer can just remove it when it's a pure functor.
I think if anything forcing it to be an instance member makes this
harder.

I looked at some other tricks to add some syntantic sugar to keep
the gt()/lt()/le()/ge() function names (either with some namespaced
inline functions or even just with macros) but it just seemed ugly.
So instead I just replaced all of the comaprisons with direct calls
to the comparator functor.