Remove "break_outer" variable
Yes, "goto considered harmful" and all that.  However, it really is the
most direct way of breaking out of multiple levels of loops.

Using a temporary boolean like this code had been doing actually requires
more work at the assembly level.  An 'if' statement that just has a
"break" or "goto" as its contents will likely compile down to a single
breanch instruction.  However, if you have to do anything else in the
body, you now have a code block that needs to be branched around in
the false case.  So in pseudo-assembly instead of:

    beq FOO

you could have:

    bne FALSE
    mov %r1, TRUE
    jmp FOO
  FALSE:

It's possible that the compiler's flowgraph optimizer will manage to
pull out the temporary variable and do this transformation itself...
but it's really asking a lot.  I don't think it's any less readable
to just change to a "goto" and express directly where we want controlflow
to go