Deactivate Valgrind build on OSX
We need a suppression file on OSX to suppress some Valgrind false
positives, but ParseAndAddCatchTests doesn't properly forward the
suppression file information so we can't make the build pass anyway.
The problem will be solved in the 2.0 version of the library thanks
to catch_discover_tests, which does it properly.

[ci skip]