Test suite maintenance
The C++98 branch will only receive minor updates if any in the future,
and there are no plans to overhaul its test suite ever. Therefore, the
branches targetting newer standards should not be held back by decisions
made for the C++98 branch, and compatibility between both can be broken
when it makes sense.

This commit replaces std::random_shuffle with a shuffle function based
on std::shuffle to be forward-compatible with standards where
std::random_shuffle has been removed. The rest is mostly on-the-fly
minor cleanups since I was working in the same area.