Simplify timsrot and timmerge signatures
Use default arguments and default template argments for the comparison
and projection functions. This reduces the number of functions to write
and to document, which will matter soon enough when new algorithms will
be added to the library.

This commit also cleans the README a bit as a drive-by fix, ensuring no
more than 120 characters per line, and adding link to cppreference when
standard library components are mentioned.